Implemented a cool test for CreateNewUserTest, and learned this:
The issue you're facing seems to be related to the way the user's ID is being auto-generated in your database.

You're using an AUTO_INCREMENT attribute for the userID field in your SQL table. When a new record is inserted into the table, MySQL automatically increments the userID value, regardless of whether old records are deleted. So, if the first user is deleted, the next user that is created will have userID equal to 2, not 1.

This behavior is causing your test to fail because you're expecting the userID to always be 1 after resetting the table, but it's actually incrementing with each new user.

To resolve this issue, you need to reset the AUTO_INCREMENT value in your resetUsersTable method. Here is the updated method:

protected static void resetUsersTable() {
    DatabaseHandler.executeCommand("DELETE FROM users;");
    DatabaseHandler.executeCommand("ALTER TABLE users AUTO_INCREMENT = 1;");
}
The command ALTER TABLE users AUTO_INCREMENT = 1; will reset the auto increment count back to 1. Now, when you create a new user after resetting the table, the userID should correctly be 1, which is what your test is expecting.